fix: runtime actions tests and flyio docker image build
