+++ python/dockerfiles/Makefile
@@ -32,9 +32,9 @@ ci: _copy-composio
 	@docker build . -t composio/composio:dev -f Dockerfile.ci
 	@rm -rf ./composio
 
-.PHONY: dev-multi
-dev-multi: _copy-composio
-	@docker buildx build --platform linux/amd64,linux/arm64 -t composio/composio:dev -f Dockerfile.dev . --push
+.PHONY: flyio
+flyio: _copy-composio
+	@docker buildx build --platform linux/amd64,linux/arm64 -t composio/composio:dev -f Dockerfile.ci . --push
 	@rm -rf ./composio
 
 .PHONY: publish
